LayoutTests/imported/w3c/web-platform-tests/fetch/sec-metadata/style.tentative.https.sub-expected.txt


FAIL Same-Origin style assert_not_equals: Empty Sec-Metadata header. got disallowed value ""
FAIL Same-Site style assert_equals: destination expected (string) "style" but got (undefined) undefined
FAIL Cross-Site style assert_equals: destination expected (string) "style" but got (undefined) undefined
